Main application finished, small adjustments to follow
Added Pseudocode to README.md which was used when planning out actual code.
Added functions takeInput, computerPlay, playRound and game that make up the entire
game along with some other helper functions that clean the body of the code up a
little. The program is now working under the five round game specified in the project
outline and will screen for correct words with both uppercase, lowercase and mixed
letters. I still have to debug for null values being entered into the takeInput
function when the player hits the cancel button, but this shouldn't be too hard. The
only other thing needing to be finished is the console log messages and potentially a
prompt asking the player if they would like to play again.