Fixed null value issue, added game looping functionality
Used a function to determine if a null value was encountered when prompted, if so the
question will be asked again prior to using the toLowerCase function on the input.
This stops the program from throwing any errors from hitting the cancel button while
being prompted. The loop game function was also added in this commit that uses a while
loop to restart the game if the player wants to after finishing a prior game This
function also calls the original prompt that asks the player if they'd like to play,
the original call is made outside the while loop. Extra console.log messages were also
added throughout the game-play to better indicate what is happening for the player and
to make the game a little more enjoyable.