Added basic in browser UI using html, css and javascript
Used the recent knowledge of event listeners and callback functions to create a basic
UI for the scissors, paper , rock game already created for the console. I have decided
to leave the 5 round functionality instead of the first to five points functionality
mentioned on TOP as I enjoyed having the round counter on the screen and there being a
fixed amount of rounds.